requirements.txt
pygame~=2.6.1
Box2D~=2.3.10
noise~=1.2.2
numpy~=2.0.2
gymnasium~=0.29.1